- (actions.c)   f_execute_command(): added length of command to error message 'command too     long'   Corrected format specifiers - (comm.c)   Corrected format specifiers   interactive_cleanup(): removed variable e   add_message(): fixed a potential crash due to a buffer overflow on the     stack using vsnprint() instead of vsprintf() (and spared the client a     comm_fatal()).   f_binary_message(): replaced a fatal() by an errorf() if a string could     not be allocated. Replaced the error message text as well, because that     is not a stack overflow (memory for the string is allocated on heap). - (xalloc.h)   Corrected format specifiers. - (object.c)   Corrected format specifiers, removed unneeded casts.   free_prog(): added refcount, program address and program name to message     for fatal() in case of negativ refcounts   v_function_exists(): changed flags to p_int.   save_array(), register_array(), register_mapping_filter(): changed i to     p_int   restore_svalue(): removed variable tmp while restoring a float.
git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2422 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/xalloc.h
@@ -16,8 +16,8 @@
 
 #define xallocate(dest,size,txt) \
     if (NULL == ((dest) = xalloc(size))) {\
-        errorf("(%s:%d) Out of memory (%lu bytes) for %s\n"\
-              , __FILE__, __LINE__, (unsigned long)(size), txt); \
+        errorf("(%s:%d) Out of memory (%zu bytes) for %s\n"\
+              , __FILE__, __LINE__, (size_t)(size), txt); \
     } else {}
 
 
@@ -41,12 +41,12 @@
     errorf("(%s:%d) Out of memory for %s\n", __FILE__, __LINE__, txt)
 
 #define outofmem(size,txt) \
-    errorf("(%s:%d) Out of memory (%lu bytes) for %s\n"\
-          , __FILE__, __LINE__, (unsigned long)(size), txt)
+    errorf("(%s:%d) Out of memory (%zu bytes) for %s\n"\
+          , __FILE__, __LINE__, (size_t)(size), txt)
 
 #define outofmem1(size,txt,arg1) \
-    errorf("(%s:%d) Out of memory (%lu bytes) for " txt "\n"\
-          , __FILE__, __LINE__, (unsigned long)(size), arg1)
+    errorf("(%s:%d) Out of memory (%zu bytes) for " txt "\n"\
+          , __FILE__, __LINE__, (size_t)(size), arg1)
 
 
 /* void memsafe(void * expr, size_t size, const char * txt)
@@ -59,8 +59,8 @@
     do { \
         size_t memsafe_size = size; \
         if (NULL == (expr)) {\
-            errorf("(%s:%d) Out of memory (%lu bytes) for %s\n"\
-                  , __FILE__, __LINE__, (unsigned long)(memsafe_size), txt); \
+            errorf("(%s:%d) Out of memory (%zu bytes) for %s\n"\
+                  , __FILE__, __LINE__, (size_t)(memsafe_size), txt); \
         } else {} \
     } while(0)
